// src/app/interfaces/register.ts
/*
  Authors : initappz (Rahul Jograna)
  Website : https://initappz.com/
  App Name : ionic 5 foodies app
  Created : 28-Feb-2021
  This App Template Source code is licensed as per the
  terms found in the Website https://initappz.com/license
  Copyright and Good Faith Purchasers © 2020-present initappz.
*/
export interface register {
    email: string;
    first_name: string;
    last_name: string;
    password: string;
    gender: string;
    mobile: string;
    fcm_token: string;
    type: string;
    lat: string;
    lng: string;
    cover: string;
    status: string;
    verified: string;
    others: string;
    date: string;
    stripe_key: string;
    cc: string;
    check: boolean;
}